+++ test/command_test.ts
@@ -5,7 +5,7 @@ import * as chassis from '../lib';
 import * as should from 'should';
 import { Client } from '@restorecommerce/grpc-client';
 import { Events, Topic } from '@restorecommerce/kafka-client';
-import * as sconfig from '@restorecommerce/service-config';
+import { createServiceConfig } from '@restorecommerce/service-config';
 import { createClient } from 'redis';
 
 
@@ -56,7 +56,7 @@ describe('CommandInterfaceService', () => {
     await validate(msg, eventName);
   };
   before(async function setup() {
-    cfg = sconfig(process.cwd() + '/test');
+    cfg = createServiceConfig(process.cwd() + '/test');
     const logger = new chassis.Logger(cfg.get('logger'));
 
     events = new Events(cfg.get('events:kafka'), logger);
@@ -347,7 +347,7 @@ describe('CommandInterfaceService', () => {
       };
       const offset = await commandTopic.$offset(-1);
       const flushCachePayload = encodeMsg({
-        data: 
+        data:
           {
             db_index: 0,
             pattern: 'user'